pyproject.toml
[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[project]
name = "my-utils"
version = "1.0.0"
description = "Shared utilities for all services"
dependencies = [
    "python-dotenv"
]

[tool.hatch.build.targets.wheel]
packages = ["utils"]
